Break L1s into multiple classes
